Change routing. Move TOS and Privacy policy outside the app
