+++ apps/frontend/src/App.tsx
@@ -2,20 +2,18 @@ import { FC, lazy, useEffect } from 'react';
 import { Router, Switch, Route } from 'react-router-dom';
 import { LastLocationProvider } from 'react-router-last-location';
 import {
-  Layout,
   BasicLayout,
   GoogleTagManager,
   ToastStack,
 } from '@asap-hub/react-components';
-import { useAuth0, useCurrentUser } from '@asap-hub/react-context';
-import { staticPages, network, welcome, logout } from '@asap-hub/routing';
+import { staticPages, welcome, logout } from '@asap-hub/routing';
 
 import history from './history';
 import CheckAuth from './auth/CheckAuth';
+import Signin from './auth/Signin';
 import Logout from './auth/Logout';
 import Frame from './structure/Frame';
 import { GTM_CONTAINER_ID } from './config';
-import { Onboardable } from './Onboardable';
 
 const loadAuthProvider = () =>
   import(/* webpackChunkName: "auth-provider" */ './auth/AuthProvider');
@@ -25,39 +23,15 @@ const loadWelcome = () =>
   import(/* webpackChunkName: "welcome" */ './welcome/Routes');
 const loadContent = () =>
   import(/* webpackChunkName: "content" */ './content/Content');
-const loadGuardedApp = () =>
-  import(/* webpackChunkName: "guarded-app" */ './GuardedApp');
+const loadAuthenticatedApp = () =>
+  import(/* webpackChunkName: "authenticated-app" */ './AuthenticatedApp');
 const Welcome = lazy(loadWelcome);
 const Content = lazy(loadContent);
-const GuardedApp = lazy(loadGuardedApp);
-
-export const ConfiguredLayout: FC = ({ children }) => {
-  const { isAuthenticated } = useAuth0();
-  const user = useCurrentUser();
-  return isAuthenticated && user ? (
-    <Onboardable>
-      {({ isOnboardable }) => (
-        <Layout
-          isOnboardable={isOnboardable}
-          userProfileHref={network({}).users({}).user({ userId: user.id }).$}
-          teams={user.teams.map(({ id, displayName = '' }) => ({
-            name: displayName,
-            href: network({}).teams({}).team({ teamId: id }).$,
-          }))}
-          aboutHref="https://www.parkinsonsroadmap.org/"
-        >
-          {children}
-        </Layout>
-      )}
-    </Onboardable>
-  ) : (
-    <BasicLayout>{children}</BasicLayout>
-  );
-};
+const AuthenticatedApp = lazy(loadAuthenticatedApp);
 
 const App: FC<Record<string, never>> = () => {
   useEffect(() => {
-    loadGuardedApp().then(loadContent).then(loadWelcome);
+    loadAuthenticatedApp().then(loadContent).then(loadWelcome);
   }, []);
 
   return (
@@ -78,27 +52,33 @@ const App: FC<Record<string, never>> = () => {
                     <Logout />
                   </Frame>
                 </Route>
-
                 <Route exact path={staticPages({}).terms.template}>
-                  <ConfiguredLayout>
+                  <BasicLayout>
                     <Frame title={null}>
                       <Content pageId="terms-and-conditions" />
                     </Frame>
-                  </ConfiguredLayout>
+                  </BasicLayout>
                 </Route>
                 <Route exact path={staticPages({}).privacyPolicy.template}>
-                  <ConfiguredLayout>
+                  <BasicLayout>
                     <Frame title={null}>
                       <Content pageId="privacy-policy" />
                     </Frame>
-                  </ConfiguredLayout>
+                  </BasicLayout>
                 </Route>
-
                 <Route>
                   <CheckAuth>
-                    <Frame title={null}>
-                      <GuardedApp />
-                    </Frame>
+                    {({ isAuthenticated }) =>
+                      !isAuthenticated ? (
+                        <Frame title={null}>
+                          <Signin />
+                        </Frame>
+                      ) : (
+                        <Frame title={null}>
+                          <AuthenticatedApp />
+                        </Frame>
+                      )
+                    }
                   </CheckAuth>
                 </Route>
               </Switch>
