+++ apps/frontend/src/auth/CheckAuth.tsx
@@ -1,24 +1,17 @@
-import { ReactNode, FC, lazy } from 'react';
+import { ReactNode, FC } from 'react';
 import { useAuth0 } from '@asap-hub/react-context';
 import { Loading } from '@asap-hub/react-components';
 
-const Signin = lazy(() => import('./Signin'));
-
 interface CheckAuthProps {
-  children: ReactNode;
+  children: (state: { isAuthenticated: boolean }) => ReactNode;
 }
 const CheckAuth: FC<CheckAuthProps> = ({ children }) => {
-  const { loading, isAuthenticated } = useAuth0();
-
+  const { isAuthenticated, loading } = useAuth0();
   if (loading) {
     return <Loading />;
   }
 
-  if (!isAuthenticated) {
-    return <Signin />;
-  }
-
-  return <>{children}</>;
+  return <>{children({ isAuthenticated: !!isAuthenticated })}</>;
 };
 
 export default CheckAuth;
